+++ packages/expo-router/src/__tests__/dismissTo.test.ios.tsx
@@ -77,7 +77,6 @@ it('should go back to a previous route in the same stack', () => {
       },
     ],
     stale: false,
-    type: 'stack',
   });
 
   act(() => router.dismissTo('/1/a'));
@@ -129,7 +128,6 @@ it('should go back to a previous route in the same stack', () => {
       },
     ],
     stale: false,
-    type: 'stack',
   });
 });
 
@@ -242,7 +240,6 @@ it('should go back to a previous route in different stacks', () => {
       },
     ],
     stale: false,
-    type: 'stack',
   });
 
   act(() => router.dismissTo('/1/a'));
@@ -294,7 +291,6 @@ it('should go back to a previous route in different stacks', () => {
       },
     ],
     stale: false,
-    type: 'stack',
   });
 });
 
@@ -401,6 +397,5 @@ it('will replace the route if the provided href is not in the history', () => {
       },
     ],
     stale: false,
-    type: 'stack',
   });
 });