+++ packages/expo-router/src/__tests__/hashs.test.ios.tsx
@@ -5,6 +5,7 @@ import { router } from '../exports';
 import { store } from '../global-state/router-store';
 import { renderRouter } from '../testing-library';
 import { parseUrlUsingCustomBase } from '../utils/url';
+import { expectCompleteStateToMatch } from './assertCompleteState';
 
 it('can push a hash url', () => {
   renderRouter({
@@ -79,7 +80,6 @@ it('can push a hash url', () => {
       },
     ],
     stale: false,
-    type: 'stack',
   });
 });
 
@@ -205,7 +205,7 @@ it('navigating to the same route with a hash will only rerender the screen', () 
     index: () => <Text testID="index" />,
   });
 
-  expect(store.state).toStrictEqual({
+  expectCompleteStateToMatch(store.state, {
     routes: [
       {
         name: '__root',
@@ -251,6 +251,5 @@ it('navigating to the same route with a hash will only rerender the screen', () 
       },
     ],
     stale: false,
-    type: 'stack',
   });
 });