+++ packages/expo-router/src/__tests__/push.test.ios.tsx
@@ -21,9 +21,25 @@ it('stacks should always push a new route', () => {
     '(group)/post/[id]/index': () => null,
   });
 
-  // Initial stale state
+  // Initial complete state
   expect(store.state).toStrictEqual({
-    routes: [{ name: '__root', state: { routes: [{ name: 'index', path: '/' }] } }],
+    index: 0,
+    key: expect.any(String),
+    routeNames: ['__root', '+not-found', '_sitemap'],
+    routes: [
+      {
+        key: expect.any(String),
+        name: '__root',
+        state: {
+          index: 0,
+          key: expect.any(String),
+          routeNames: ['index', '(group)'],
+          routes: [{ key: expect.any(String), name: 'index', path: '/' }],
+          stale: false,
+        },
+      },
+    ],
+    stale: false,
   });
 
   /**
@@ -187,7 +203,6 @@ it('stacks should always push a new route', () => {
       },
     ],
     stale: false,
-    type: 'stack',
   });
 });
 
@@ -383,7 +398,6 @@ it('works in a nested layout Stack->Tab->Stack', () => {
       },
     ],
     stale: false,
-    type: 'stack',
   });
 });
 
@@ -514,7 +528,6 @@ it('targets the correct Stack when pushing to a nested layout', () => {
       },
     ],
     stale: false,
-    type: 'stack',
   });
 });
 
@@ -532,21 +545,31 @@ it('push should also add anchor routes', () => {
     '(group)/orange': () => null,
   });
 
-  // Initial stale state
+  // Initial complete state
   expect(store.state).toStrictEqual({
+    index: 0,
+    key: expect.any(String),
+    routeNames: ['__root', '+not-found', '_sitemap'],
     routes: [
       {
+        key: expect.any(String),
         name: '__root',
         state: {
+          index: 0,
+          key: expect.any(String),
+          routeNames: ['index', '(group)'],
           routes: [
             {
+              key: expect.any(String),
               name: 'index',
               path: '/',
             },
           ],
+          stale: false,
         },
       },
     ],
+    stale: false,
   });
 
   act(() => router.push('/orange', { withAnchor: true }));
@@ -602,7 +625,6 @@ it('push should also add anchor routes', () => {
       },
     ],
     stale: false,
-    type: 'stack',
   });
 });
 
@@ -676,7 +698,6 @@ describe('singular', () => {
         },
       ],
       stale: false,
-      type: 'stack',
     });
 
     // Should push /apple and remove all previous instances of /apple
@@ -721,7 +742,6 @@ describe('singular', () => {
         },
       ],
       stale: false,
-      type: 'stack',
     });
   });
 
@@ -806,7 +826,6 @@ describe('singular', () => {
         },
       ],
       stale: false,
-      type: 'stack',
     });
 
     // Should push /apple and remove all previous instances of /apple
@@ -855,7 +874,6 @@ describe('singular', () => {
         },
       ],
       stale: false,
-      type: 'stack',
     });
   });
 });